big update:
there was a bug when R_bias=1.0 caused by sampling of the halos.
This happened because hmf is nan for values larger than the virial
mass of the region. However, mass function cleaning part only
caught zeros, not nans. Now I've added np.isnan to capture any nans
in the mass function.